complete/backend/requirements.txt
fastapi>=0.103.1
uvicorn>=0.22.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pytest>=7.0.0
httpx>=0.24.0
pydantic-settings>=2.0.0